Fix some unit tests
